WIP on updating duplicate publication group workflows
This implements the new workflow that we need for merging publications
and deleting duplicate groups, but it’s missing some integration tests,
and some of the old integration tests haven’t been updated so they’re
still failing. Also included is a stub for the “ignore false duplicates”
workflow, but this workflow hasn’t been implemented yet.

+++ app/controllers/custom_admin/publication_merges_controller.rb
@@ -1,26 +1,39 @@
 class CustomAdmin::PublicationMergesController < RailsAdmin::ApplicationController
   def create
     group = DuplicatePublicationGroup.find(params[:duplicate_publication_group_id])
-    selected_pub = Publication.find(params[:publication_id])
-    pubs_to_delete = group.publications - [selected_pub]
 
-    ActiveRecord::Base.transaction do
-      imports = group.publications.map { |p| p.imports }.flatten
-
-      imports.each do |i|
-        i.update_attributes!(publication: selected_pub)
+    if params[:commit] == 'Merge Selected'
+      if params[:merge_target_publication_id].blank? ||
+         params[:selected_publication_ids].blank? ||
+          params[:selected_publication_ids] == [params[:merge_target_publication_id]]
+        flash[:error] = "To perform a merge, you must select a merge target and at least one other publication to merge into the target."
+        redirect_to rails_admin.show_path(model_name: :duplicate_publication_group, id: group.id)
+        return
       end
 
-      pubs_to_delete.each do |p|
-        p.destroy
+      merge_target_pub = group.publications.find(params[:merge_target_publication_id])
+      selected_pubs = group.publications.find(params[:selected_publication_ids])
+      pubs_to_delete = selected_pubs - [merge_target_pub]
+
+      ActiveRecord::Base.transaction do
+        imports_to_reassign = pubs_to_delete.map { |p| p.imports }.flatten
+
+        imports_to_reassign.each do |i|
+          i.update_attributes!(publication: merge_target_pub)
+        end
+
+        pubs_to_delete.each do |p|
+          p.destroy
+        end
+
+        merge_target_pub.update_attributes!(updated_by_user_at: Time.current)
       end
 
-      selected_pub.update_attributes!(duplicate_group: nil,
-                                      updated_by_user_at: Time.current)
-      group.destroy
+      flash[:success] = I18n.t('admin.publication_merges.create.success')
+    elsif params[:commit] = 'Ignore Selected'
+      flash[:alert] = "This feature has not been implemented yet."
     end
 
-    flash[:success] = I18n.t('admin.publication_merges.create.success')
-    redirect_to rails_admin.edit_path(model_name: :publication, id: selected_pub.id)
+    redirect_to rails_admin.show_path(model_name: :duplicate_publication_group, id: group.id)
   end
 end
